requirements.txt
# Data collection
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.1.0
kaggle>=1.6.0

# ML models
scikit-learn>=1.4.0
xgboost>=2.0.0
lightgbm>=4.0.0
numpy>=1.26.0

# LLM agents
anthropic>=0.40.0

# Backend API
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.5.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
